Apply to Moodle4.5
